Added HTB Lame, HTB Legacy articles; Added HTB template for future articles; Updated htb-active and pelicanconf.py
Added new articles for Hack the Box Lame and Legacy.
Added a template for future Hack the Box articles.
Added an exclusion for the file in pelicanconf.py so the template doesn't generate an article.
Added a tag for the Active article so it will be in the correct categories.